Heat Transfer

Heat transfer equations

Here is a short review of heat transfer. There are only a few equations that drive this process. For all equations below $A$ is area, $L$ is length, $T$ is temperature in Kelvin.

Conduction

Conduction is when heat moves through physical contact with objects. In the sauna, the main conductive aspect is the walls conducting temperature to the outside. Effective insulation will greatly reduce this mode of heat transfer.

\[k\frac{(T_{hot}-T_{cold})A}{L}\]

where $k$ is the conduction coefficient, which is fairly straight forward to measure for many materials. To learn all you probably need to know about thermal conduction consult Wikipedia.. In SaunaModel the only conduction computed is through the walls of the sauna.

SaunaModel.conduction_exchange_wall — Function.

conduction_exchange_wall(temperature_room::Temperature, temperature_outside::Temperature, room::Room)::Power Provides an estimate of heat loss through the walls of the sauna.

source

Convection

Convection is the mode of heat transfer due to fluid flows. This mode of heat transfer is extremely common in a sauna. The heat transfer from a burning hot steam to your shoulders is one. Heat transfer from hot gas rising off the fire to the stove is another.

\[h(T_{hot}-T_{cold})A\]

where $h$ is the convective coefficient. Convective coefficients can be approximated through a fairly complex process of determining Reynold's numbers and Nusselt's numbers. I used reference numbers to save time. To learn all you probably need to know about thermal convection consult Wikipedia. In SaunaModel convection exchange is computed using convection_exchange.

SaunaModel.convection_exchange — Function.

convection_exchange(temperature_1::Temperature, temperature_2::Temperature, area::Area, convective_coeff )::Power Provides an estimate of convective power exchange on an area. Convection coefficient estimation is an important and difficult part of this.

source

Radiation

Radiation is the mode of heat transfer due to electromagnetic radiation. Heat moving like light, instantaneously, through the atmosphere. If you feel the heat coming off of something hot, it is probably radiation. The burning sensation in your shins from an oversized stove (cough, Benda's, cough) is from radiation.

\[\sigma\epsilon A (T_{hot}^4 - T_{cold}^4)\]

where $\sigma$ is the Stefan–Boltzmann constant (some universal constant), $\epsilon$ is the emissivity (black top has a high emissivity, white shirts have a lower one). I assumed that $\epsilon=1$ for the purposes of this model. Wikipedia. In SaunaModel, radiance exchange is computed using radiance_exchange.

SaunaModel.radiance_exchange — Function.

radiance_exchange(temperature_1::Temperature, temperature_2::Temperature, area::Area, emissivity=1)::Power Provides an estimate of the raidance power transmitted between two heat sources with a certain area exposure.

source

Advection (mass transfer)

Heat is also transfered when hot air leaves the room and cold air flows in. Since our sauna is not leak proof, a certain amount of the air in the room turns over on every time increment. This rate is associated with the difference in air temperature. Nature had a fairly solid article on the turnover rate for a house, and I adjusted (upwards) it for the surface area to volume ratio of the sauna under consideration. All smaller rooms have a larger surface area to volume ratio compared to larger rooms. The mass exchange between the sauna and the outdoors is estimated in air_turnover.

SaunaModel.air_turnover — Function.

air_turnover(temperature_difference::Temperature, room::Room)::Frequency A house was shown to have a turnover rate of .176 + .0162 ΔT, and estimates of the house surface area and volume could be made. This allows for an estimate of air turnover in a sauna. https://www.nature.com/articles/7500229

source

Phase change

Throwing water on the stove causes heat to convect into the water on the stove rapidly turning it into steam. This phase change heats the air and stings your ears. Since this is also a mass transfer, it is also advection. The steam coming off the stove is assumed to be 212°F as there will be minimal heating of the steam after it boils.