API

SaunaModel API

The sauna model api is fairly straight forward to use. An example json to submit as a job can be obtained by calling the get_default_scenario_json function. For example

SaunaModel.get_default_scenario_json — Function.

Returns a default json configuration –for example "initial_temperature_air":{"val":51,"unit":"°F"}

source

Use these defaults as a guide to use the dictionary_api – only one value, of the unit that was returned in get_default_scenario_json must be submitted.

SaunaModel.dictionary_api — Function.

Submit job description here, get a dictionary of values back. Temperatures are in Farenheit, humidity is in Pa, and power is in Watts.

source