Guide

Using SaunaModel

To use the sauna model, the first thing to be done is to contruct your model of the sauna. Consulting the default elements. This can be a bit exahustive to do from scratch.

SaunaModel.Stove — Type.

A stove can be described by its dimensions and specific heats. The stove is assumed to be made out of steel.

source

SaunaModel.Room — Type.

A room is described by its dimensions, heat transfer coefficients, and specific heat. A density is assumed to compute room mass.

source

SaunaModel.SaunaNoWater — Type.

A sauna consists of a stove and room description, as well as a view factor of the room itself from the sauna, which is an estimate of how much of the stoves area is exposed to the room.

source

SaunaModel.SteamThrowing — Type.

Steam throwing describes how steam is thrown in a sauna. A temperature to start throwing steam must be selected in order to know when someone would enter the sauna.

source

SaunaModel.Fire — Type.

The fire is the driving heat source of the sauna system. A sigmoid heating function is used to drive the fire temperature up over about a 20 minute period

source

SaunaModel.SaunaScenario — Type.

The sauna scenario brings many of the components together, as well as initial and boudnary conditions.

source

Then, you will want submit your solution to solve_sauna, where the differential equations magic happens.

SaunaModel.solve_sauna — Function.

This function sets up the differential equation solvers, including the jump solver, runs the solver, then maps units back onto the results and returns a SaunaResults

source

It will then return a SaunaResults which can be parsed.

SaunaModel.SaunaResults — Type.

Sauna results store the state of the system at each time step. Temperatures of the stove, air, room, and water on stove are tracked as well as room humidity and mass of water on stove.

source

Defaults

It takes quite a bit of legwork to track down all the correct parameters for a sauna. I kept some reasonable parameters in a file under src/api/default_components.jl. I would reccomend starting there if you want to edit a sauna from the commmand line.